# Improve performance for truncating tables in tests